+++ lib/src/service/sender/smtp_mail_sender_service.dart
@@ -1,4 +1,4 @@
-import 'package:impaktfull_mailer/impaktfull_mailer.dart';
+import 'package:impaktfull_mailer/src/service/sender/mail_sender_service.dart';
 import 'package:mailer/smtp_server.dart';
 
 abstract class SmtpMailSenderService extends MailSenderService {